% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{ora}
\alias{ora}
\title{Fisher's exact test for pathway enrichment}
\usage{
ora(query_set, universe_entrez, universe_symbols, pathway_matrix)
}
\arguments{
\item{query_set}{Gene signature of interest, as EntrezIDs}

\item{universe_entrez}{Full EntrezID set where the query set comes from (usually the full set of RNA-seq or microarray genes characterized in the transcriptomics platform)}

\item{universe_symbols}{Gene symbols for the entire set of genes in universe}

\item{pathway_matrix}{A (sparse) matrix where each row is a pathway and each column a gene. As distributed, this matrix has pathway sets available in CPDB and MSigDB.}
}
\value{
A tibble with overlap of query set with pathway genes, set of query genes in pathway (when found), and Fisher's exact test p-value
}
\description{
\code{ora} returns the enrichment of pathway sets given a query gene signature based on Fisher's exact test. Main function in the \code{\link{oranges}} package.
}
