% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oranges.R
\name{oranges}
\alias{oranges}
\title{ORANGES: Over-Representation ANalysis for Gene Expression Signatures}
\usage{
oranges(query_entrez, universe_entrez)
}
\arguments{
\item{universe_entrez}{Full EntrezID set where the query set comes from (usually the full set of RNA-seq or microarray genes characterized in the transcriptomics platform)}

\item{query_set}{Gene signature of interest, as EntrezIDs}

\item{universe_symbols}{Gene symbols for the entire set of genes in universe}
}
\value{
A tibble.
}
\description{
\code{oranges} returns the enrichment of pathway sets given a query gene signature based on Fisher's exact test. Currently this function is built around enrichment of human pathway sets, but \code{\link{ora}} function can be generalized to fit different organisms of interest provided a pathway matrix is provided.
}
