% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oranges.R
\name{plot_oranges}
\alias{plot_oranges}
\title{Plot enrichment results}
\usage{
plot_oranges(res, num_pathways)
}
\arguments{
\item{res}{ORANGES results data frame}

\item{thr}{Significance (FDR-corrected q-value) for pathway enrichment.  Defaults to q < 0.05}
}
\description{
\code{plot_oranges} returns a plot for the top 25 enriched pathways identified with the \code{\link{oranges}} package.
}
